Add card score module
